move example to test folder and add more test cases
